move DB population to another class
