+++ src/main/java/com/jarqprog/artGallery/springData/dataLoader/InitialDataLoader.java
@@ -1,0 +1,6 @@
+package com.jarqprog.artGallery.springData.dataLoader;
+
+public interface InitialDataLoader {
+
+    void populateDb();
+}
